+++ SocialCasino/Models/User.cs
@@ -8,7 +8,7 @@ public class User
     public string LastName { get; set; }
     public string Email { get; set; }
     public string Password { get; set; }
-    public uint Balance { get; set; } = 0;
+    public decimal Balance { get; set; } = 0;
     public bool IsActive { get; set; } = true;
     public DateTime CreatedAt { get; set; } = DateTime.Now;
 }
